Match Input to known list of tags than Ajax
Send javascript an array of known media tags so that we can match that
to what is entered into the input box.  Then we send the request for
media tag that has been matched back to server and respond with list of
media items.
Oh yeah and I forgot to add the media tags icon to the mix on the last
commit so I added that.