Fixes for check-all and search icons
I have no idea what I was thinkin but I sorted out the the variable
issue that prevented check-all from working.  I also realized that
clicking on the tag icon in the selected media popup opened up the
autocomplete menu on the input box with unexpected results.  I properly
identified the element that should be targeted for the different search
functions.