New total and viewable counts plus Grid label
Re-work of total and viewable items.  Moved them to under search box.
Trying to have a quick, easy way to show all items.  Total item count is
now linked to wp_query search with no pagination.  Not perfect but
close.  Also wanted to give feedback as to what search term was used so
Grid is now labeled with tag searched.